complete backend

+++ backend/controller/goalcontroller.js
@@ -1,11 +1,12 @@
 const asynchandler=require('express-async-handler')
 const Goal=require("../models/goalmodel")
+const User=require('../models/usermodel')
 // @desc  get goals
 // @routes get  api/getgoals
 // @access private
 const getgoals= asynchandler(async(req,res)=>
 {
-    const goals=await Goal.find()
+    const goals=await Goal.find({user:req.user.id})
     res.status(200).json(goals)
 })
 
@@ -20,7 +21,10 @@ const setgoals= asynchandler(async(req,res)=>
         res.status(400)
         throw new Error('Please add the text!!')
     }
-    const goals=await Goal.create({text:req.body.text})
+    const goals=await Goal.create({
+        text:req.body.text,
+        user:req.user.id
+    })
     res.json(goals)
 })
 
@@ -34,6 +38,17 @@ const updategoals= asynchandler(async(req,res)=>
     {
         res.status(400)
         throw new Error("ID not found")
+    }
+    const user=await User.findById(req.user.id)
+    if(!user)
+    {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    if(goal.user.id.toString()!==user.id)
+    {
+        res.status(401)
+        throw new Error('User not Authorized')
     }
     const updatedgoal=await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
     res.status(200).json(updatedgoal)
@@ -50,6 +65,17 @@ const deletegoals= asynchandler(async(req,res)=>
     res.status(400)
     throw new Error("Id not found")
    }
+   const user=await User.findById(req.user.id)
+    if(!user)
+    {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    if(goal.user.id.toString()!==user.id)
+    {
+        res.status(401)
+        throw new Error('User not Authorized')
+    }
    await Goal.findByIdAndDelete(req.params.id)
    res.status(200).json({id:req.params.id})
 })
